implement room status views.

+++ app/src/main/java/chat/rocket/android/content/subscriber/StreamNotifyRoom.java
@@ -9,7 +9,6 @@ import org.json.JSONArray;
 import org.json.JSONException;
 import org.json.JSONObject;
 
-import chat.rocket.android.api.JSONParseEngine;
 import chat.rocket.android.api.ws.RocketChatWSAPI;
 import chat.rocket.android.content.RocketChatDatabaseHelper;
 import chat.rocket.android.model.Message;
@@ -17,11 +16,8 @@ import jp.co.crowdworks.android_ddp.ddp.DDPSubscription;
 
 public class StreamNotifyRoom extends AbstractRocketChatSubscription {
 
-    private final JSONParseEngine mParser;
-
     public StreamNotifyRoom(Context context, Looper looper, RocketChatWSAPI api) {
         super(context, looper, api);
-        mParser = new JSONParseEngine(context);
     }
 
     @Override
@@ -35,20 +31,22 @@ public class StreamNotifyRoom extends AbstractRocketChatSubscription {
     }
 
     @Override
-    protected void onDocumentAdded(DDPSubscription.DocEvent docEvent) throws JSONException {
-        final JSONArray args = ((DDPSubscription.Added) docEvent).fields.getJSONArray("args");
-        final String path = args.getString(0);
-        final JSONObject obj = args.getJSONObject(1);
+    protected void onDocumentAdded(DDPSubscription.Added docEvent) throws JSONException {
+        if(!docEvent.fields.isNull("args")) {
+            final JSONArray args = docEvent.fields.getJSONArray("args");
+            final String path = args.getString(0);
+            if (!TextUtils.isEmpty(path) && path.endsWith("/deleteMessage")) {
+                final JSONObject obj = args.getJSONObject(1);
 
-        if(!TextUtils.isEmpty(path) && path.endsWith("/deleteMessage")) {
-            final String messageID = obj.getString("_id");
-            Message m = RocketChatDatabaseHelper.read(mContext, new RocketChatDatabaseHelper.DBCallback<Message>() {
-                @Override
-                public Message process(SQLiteDatabase db) throws Exception {
-                    return Message.getById(db, messageID);
-                }
-            });
-            if(m!=null) m.deleteByContentProvider(mContext);
+                final String messageID = obj.getString("_id");
+                Message m = RocketChatDatabaseHelper.read(mContext, new RocketChatDatabaseHelper.DBCallback<Message>() {
+                    @Override
+                    public Message process(SQLiteDatabase db) throws Exception {
+                        return Message.getById(db, messageID);
+                    }
+                });
+                if (m != null) m.deleteByContentProvider(mContext);
+            }
         }
     }
 }
